어댑터 추상화, 사용하지 않는 클래스 삭제
