+++ app/src/main/java/com/boostcamp/travery/OnItemClickListener.kt
@@ -1,0 +1,5 @@
+package com.boostcamp.travery
+
+interface OnItemClickListener {
+    fun onItemClick(item: Any)
+}
