--- app/src/main/java/com/boostcamp/travery/main/adapter/GroupItem.kt
@@ -1,3 +1,0 @@
-package com.boostcamp.travery.main.adapter
-
-data class GroupItem(val group: String)
