--- app/src/main/java/com/boostcamp/travery/main/adapter/OnItemClickListener.kt
@@ -1,7 +1,0 @@
-package com.boostcamp.travery.main.adapter
-
-import com.boostcamp.travery.data.model.Route
-
-interface OnItemClickListener {
-    fun onItemClick(item: Route)
-}
