--- app/src/main/res/layout/item_route_list_group.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:tools="http://schemas.android.com/tools"
-		xmlns:android="http://schemas.android.com/apk/res/android">
-
-	<data>
-		<variable
-				name="item"
-				type="com.boostcamp.travery.main.adapter.GroupItem"/>
-	</data>
-
-	<LinearLayout
-			android:layout_width="match_parent"
-			android:layout_height="match_parent"
-			android:padding="10dp">
-
-		<TextView
-				tools:text="오늘"
-				android:text="@{item.group}"
-				android:textSize="20sp"
-				android:layout_width="wrap_content"
-				android:layout_height="wrap_content"/>
-	</LinearLayout>
-</layout>
